+++ src/main/java/acessivel/entity/Queixa.java
@@ -20,7 +20,7 @@ public class Queixa extends Solicitacao {
 
     @Id
     @Column(name = "id_queixa")
-    private int id_queixa;
+    private Long id_queixa;
 
     @Column(name = "imagem_link")
     private String imagem_link;
